Move some responsibilities in Torchx
